requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==21.3
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
telegram==0.0.1
tornado==6.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
